refactor: modularize cli command groups
