+++ src/cli/types.ts
@@ -1,0 +1,81 @@
+export type CliOptions = {
+  json?: boolean;
+  yes?: boolean;
+  assist?: boolean;
+  force?: boolean;
+  all?: boolean;
+  interval?: string;
+  timeout?: string;
+  concurrency?: string;
+  continueOnError?: boolean;
+  noUi?: boolean;
+  dryRun?: boolean;
+  apply?: boolean;
+  dangerous?: boolean;
+  network?: boolean;
+  printCd?: boolean;
+  openShell?: boolean;
+  exec?: string;
+  quiet?: boolean;
+  verbose?: boolean;
+  debug?: boolean;
+  trace?: boolean;
+  task?: string;
+  since?: string;
+  until?: string;
+  provider?: string;
+  repo?: string;
+  answer?: string | string[];
+  planOnly?: boolean;
+  priority?: string;
+  queue?: boolean;
+  githubToken?: string;
+  linearToken?: string;
+  model?: string;
+  modelPlan?: string;
+  modelExecute?: string;
+  modelReview?: string;
+  modelVerify?: string;
+  modelPr?: string;
+  modelRecovery?: string;
+  cognitionProvider?: string;
+  cognitionModelKickoff?: string;
+  cognitionModelPlan?: string;
+  cognitionModelReview?: string;
+  cognitionModelCi?: string;
+  cognitionModelVerify?: string;
+  cognitionModelRecovery?: string;
+  cognitionModelPr?: string;
+  cognitionModelConversationSummary?: string;
+  maxTurns?: string;
+  maxTurnsPlan?: string;
+  maxTurnsExecute?: string;
+  maxTurnsReview?: string;
+  maxTurnsVerify?: string;
+  maxTurnsPr?: string;
+  maxTurnsRecovery?: string;
+  maxBudgetUsd?: string;
+  maxBudgetUsdPlan?: string;
+  maxBudgetUsdExecute?: string;
+  maxBudgetUsdReview?: string;
+  maxBudgetUsdVerify?: string;
+  maxBudgetUsdPr?: string;
+  maxBudgetUsdRecovery?: string;
+  maxThinkingTokens?: string;
+  maxThinkingTokensPlan?: string;
+  maxThinkingTokensExecute?: string;
+  maxThinkingTokensReview?: string;
+  maxThinkingTokensVerify?: string;
+  maxThinkingTokensPr?: string;
+  maxThinkingTokensRecovery?: string;
+  maxToolCalls?: string;
+  maxToolMs?: string;
+  maxReviewLoops?: string;
+  persistSessions?: boolean;
+  verifyShell?: string;
+  stateMode?: string;
+  title?: string;
+  desc?: string;
+  ac?: string | string[];
+  fromFile?: string;
+};
